conditionally start env

+++ start-container.sh
@@ -9,7 +9,7 @@
 ## Description :
 ## --
 ## Created : <2017-08-28>
-## Updated: Time-stamp: <2018-06-27 21:03:25>
+## Updated: Time-stamp: <2018-06-27 21:17:37>
 ##-------------------------------------------------------------------
 case $LANG_DEV in
     golang)
@@ -27,10 +27,12 @@ case $LANG_DEV in
         exit 1
 esac
 
-echo "Stop existing container($CONTAINER_NAME), if running."
-set -x
-docker stop "$CONTAINER_NAME"; docker rm "$CONTAINER_NAME" || true
+if docker ps | grep $CONTAINER_NAME 1>/dev/null 2>&1 ; then
+    echo "Stop existing container($CONTAINER_NAME), if running."
+    docker stop "$CONTAINER_NAME"; docker rm "$CONTAINER_NAME" || true
+fi
 
+set -x
 docker run -t -d -v ${SRC_DIR}:/root/code \
        -h "$CONTAINER_NAME" --name "$CONTAINER_NAME" --entrypoint=/bin/sh -w /root/code "$DOCKER_IMAGE"
 set +x
